Add simulation result output and verify results match reference impl
I added the code that enables the simulator to write a JSON-like format
containing information about the car time, speed, location.

I then compared the output of the C implementation with the Python implementation.
The only difference is that Python will round floating point values, whereas the
C implementation does not.  This is likely not an issue, since the output is
used to generate images/videos.  Since the underlying values are the same, the
result will not change.