Config -> JSON, drop/lie comm strategies
I have refactored the config to be in JSON rather than .ini, because
the data was heirarchal in the sense that a guide_strategy has
information that is related to it.  If everything was on the same
level, it would be hard to access information initialization info.

I also created two new communication strategies:
- random drop: randomly drops a percentage of packets (default=.1)
- random lie: randomly lies about the value of the acceleration.

I also set it so that the output of the simulation is written to a
directory.  This helps manage the files, because the directory has
been ignored by git.