Add Sensor Error with bias and standard deviation
The sensors can now be wrong by having the value dislocated by
a set bias and/or distorted according to a gaussian std.

Control mode decisions are made based on this information, which
means that the collision detection had to change a little.  Under
the assumption that a car would probably be able to physically sense
that it has crashed (rather than relying on navigation systems), I
moved the crash detection out of the guidance, and into the world
object.  After all updates are performed, the world checks that none
of the cars crashed.  The actual global positions are checked rather
than the distorted positions from the sensor.

The simulator can also do multiple runs to calculate the number of
crashes.  The only aggregation right now is the number of crashes,
so you don't get pretty charts out of it.